# openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "ProyectoAtlas.Api | v1",
    "version": "1.0.0"
  },
  "paths": {
    "/projects/{projectSlug}/documentations/{documentationSlug}/activity-feed": {
      "get": {
        "tags": [
          "DocumentationActivityFeed"
        ],
        "parameters": [
          {
            "name": "projectSlug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "documentationSlug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetDocumentationActivityFeedResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/projects/{projectSlug}/milestones/{milestoneSlug}/features": {
      "post": {
        "tags": [
          "MilestoneFeatureLinks"
        ],
        "parameters": [
          {
            "name": "projectSlug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "milestoneSlug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMilestoneFeatureLinkCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMilestoneFeatureLinkCommand"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMilestoneFeatureLinkCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MilestoneFeatureLink"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "MilestoneFeatureLinks"
        ],
        "parameters": [
          {
            "name": "projectSlug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "milestoneSlug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListMilestoneFeatureLinksResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/projects/{projectSlug}/milestones/{milestoneSlug}/features/{linkId}": {
      "delete": {
        "tags": [
          "MilestoneFeatureLinks"
        ],
        "parameters": [
          {
            "name": "projectSlug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "milestoneSlug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "linkId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "204": {
            "description": "No Content",
            "content": {
              "application/json": { }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/projects/{projectSlug}/features/{slug}": {
      "get": {
        "tags": [
          "ProjectFeatures"
        ],
        "parameters": [
          {
            "name": "projectSlug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Feature"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "ProjectFeatures"
        ],
        "parameters": [
          {
            "name": "projectSlug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProjectFeatureCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProjectFeatureCommand"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProjectFeatureCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Feature"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "ProjectFeatures"
        ],
        "parameters": [
          {
            "name": "projectSlug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "204": {
            "description": "No Content",
            "content": {
              "application/json": { }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/projects/{projectSlug}/milestones/{slug}": {
      "get": {
        "tags": [
          "ProjectMilestones"
        ],
        "parameters": [
          {
            "name": "projectSlug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Milestone"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "ProjectMilestones"
        ],
        "parameters": [
          {
            "name": "projectSlug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProjectMilestoneCommand"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProjectMilestoneCommand"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProjectMilestoneCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Milestone"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "ProjectMilestones"
        ],
        "parameters": [
          {
            "name": "projectSlug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "204": {
            "description": "No Content",
            "content": {
              "application/json": { }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiErrorResponse": {
        "required": [
          "statusCode",
          "message",
          "code"
        ],
        "type": "object",
        "properties": {
          "statusCode": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "description": "HTTP status code returned by the API.",
            "format": "int32"
          },
          "message": {
            "type": "string",
            "description": "Developer-oriented message that explains the error."
          },
          "code": {
            "type": "string",
            "description": "Stable application error code intended for frontend flows."
          }
        },
        "example": {
          "statusCode": 409,
          "message": "Project slug 'proyecto-atlas' already exists.",
          "code": "PROJECT_SLUG_CONFLICT"
        }
      },
      "AuditAction": {
        "enum": [
          "Created",
          "Updated",
          "Deleted"
        ],
        "type": "string"
      },
...
      "CreateMilestoneFeatureLinkCommand": {
        "required": [
          "featureSlug"
        ],
        "type": "object",
        "properties": {
          "featureSlug": {
            "type": "string"
          }
        },
        "example": {
          "featureSlug": "authentication-api"
        }
      },
...
      "DocumentationActivityFeedItem": {
        "required": [
          "type",
          "occurredAtUtc"
        ],
        "type": "object",
        "properties": {
          "type": {
            "$ref": "#/components/schemas/DocumentationActivityFeedItemType"
          },
          "occurredAtUtc": {
            "type": "string",
            "format": "date-time"
          },
          "auditAction": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/AuditAction"
              }
            ]
          },
          "entitySlug": {
            "type": [
              "null",
              "string"
            ]
          },
          "entityTitle": {
            "type": [
              "null",
              "string"
            ]
          },
          "versionNumber": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "versionStatus": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/DocumentationStatus"
              }
            ]
          },
          "relationKind": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/DocumentationRelationKind"
              }
            ]
          },
          "relationDirection": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/DocumentationActivityRelationDirection"
              }
            ]
          },
          "relatedDocumentationSlug": {
            "type": [
              "null",
              "string"
            ]
          },
          "relatedDocumentationTitle": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "DocumentationActivityFeedItemType": {
        "enum": [
          "AuditEvent",
          "VersionCreated",
          "RelationCreated"
        ],
        "type": "string"
      },
      "DocumentationActivityRelationDirection": {
        "enum": [
          "Incoming",
          "Outgoing"
        ],
        "type": "string"
      },
...
      "DocumentationRelationKind": {
        "enum": [
          "RelatedTo",
          "DependsOn",
          "Implements",
          "Supersedes"
        ],
        "type": "string"
      },
...
      "DocumentationStatus": {
        "enum": [
          "Draft",
          "Published",
          "Archived"
        ],
        "type": "string"
      },
...
      "Feature": {
        "required": [
          "projectId",
          "title",
          "summary",
          "status"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "projectId": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "summary": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/FeatureStatus"
          },
          "createdAtUtc": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAtUtc": {
            "type": "string",
            "format": "date-time"
          }
        },
        "example": {
          "id": "b9a57f81-8729-4f58-a65a-4f0b7b8fc11d",
          "projectId": "8b658c72-8f6f-4fef-9d65-f2fa6eb60bd7",
          "title": "Authentication API",
          "summary": "Expose the first auth endpoints for Atlas clients.",
          "slug": "authentication-api",
          "status": "Planned",
          "createdAtUtc": "2026-04-01T12:00:00Z",
          "updatedAtUtc": "2026-04-01T12:00:00Z"
        }
      },
...
      "FeatureStatus": {
        "enum": [
          "Planned",
          "InProgress",
          "Done"
        ],
        "type": "string"
      },
      "GetDocumentationActivityFeedResponse": {
        "required": [
          "items"
        ],
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DocumentationActivityFeedItem"
            }
          }
        }
      },
...
      "ListMilestoneFeatureLinksResponse": {
        "required": [
          "items"
        ],
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MilestoneFeatureLink"
            }
          }
        },
        "example": {
          "items": [
            {
              "id": "a1a57f81-8729-4f58-a65a-4f0b7b8fc11d",
              "projectId": "8b658c72-8f6f-4fef-9d65-f2fa6eb60bd7",
              "milestoneId": "e39bc9ba-4a39-48f8-9220-5af3eb68c737",
              "featureId": "b9a57f81-8729-4f58-a65a-4f0b7b8fc11d",
              "createdAtUtc": "2026-04-24T12:00:00Z"
            }
          ]
        }
      },
...
      "Milestone": {
        "required": [
          "projectId",
          "title",
          "summary",
          "status"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "projectId": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "summary": {
            "type": "string"
          },
          "slug": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/MilestoneStatus"
          },
          "targetDateUtc": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "createdAtUtc": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAtUtc": {
            "type": "string",
            "format": "date-time"
          }
        },
        "example": {
          "id": "e39bc9ba-4a39-48f8-9220-5af3eb68c737",
          "projectId": "8b658c72-8f6f-4fef-9d65-f2fa6eb60bd7",
          "title": "MVP Release",
          "summary": "Cerrar la primera entrega publica del proyecto.",
          "slug": "mvp-release",
          "status": "Planned",
          "targetDateUtc": "2026-05-15T00:00:00Z",
          "createdAtUtc": "2026-04-08T12:00:00Z",
          "updatedAtUtc": "2026-04-08T12:00:00Z"
        }
      },
      "MilestoneFeatureLink": {
        "required": [
          "projectId",
          "milestoneId",
          "featureId"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "projectId": {
            "type": "string",
            "format": "uuid"
          },
          "milestoneId": {
            "type": "string",
            "format": "uuid"
          },
          "featureId": {
            "type": "string",
            "format": "uuid"
          },
          "createdAtUtc": {
            "type": "string",
            "format": "date-time"
          }
        },
        "example": {
          "id": "a1a57f81-8729-4f58-a65a-4f0b7b8fc11d",
          "projectId": "8b658c72-8f6f-4fef-9d65-f2fa6eb60bd7",
          "milestoneId": "e39bc9ba-4a39-48f8-9220-5af3eb68c737",
          "featureId": "b9a57f81-8729-4f58-a65a-4f0b7b8fc11d",
          "createdAtUtc": "2026-04-24T12:00:00Z"
        }
      },
      "MilestoneStatus": {
        "enum": [
          "Planned",
          "InProgress",
          "Done"
        ],
        "type": "string"
      },
...
      "UpdateProjectFeatureCommand": {
        "type": "object",
        "properties": {
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "summary": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/FeatureStatus"
              }
            ]
          }
        },
        "example": {
          "title": "Authentication API",
          "summary": "Expose login and token refresh endpoints.",
          "status": "InProgress"
        }
      },
...
      "UpdateProjectMilestoneCommand": {
        "required": [
          "title",
          "summary",
          "status"
        ],
        "type": "object",
        "properties": {
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "summary": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/MilestoneStatus"
              }
            ]
          },
          "targetDateUtc": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "clearTargetDate": {
            "type": "boolean",
            "default": false
          }
        },
        "example": {
          "title": "MVP Release",
          "summary": "Cerrar la primera entrega publica con documentacion y features base.",
          "status": "InProgress",
          "targetDateUtc": "2026-05-20T00:00:00Z",
          "clearTargetDate": false
        }
      },
...
    }
  },
...
}